Create separate route + controller for set
The problem with splitting the route is that the app memory is no
longer shared within the routes as it was when everything was
contained in one file.

To share the app memory to routes on separate files, we create the
`app.locals.memory` object, which is shared with the app and can be
accessed within any route as `req.app.locals.memory`.